+++ app/src/main/java/org/simple/clinic/teleconsultlog/prescription/TeleconsultPrescriptionUiActions.kt
@@ -6,4 +6,5 @@ interface TeleconsultPrescriptionUiActions {
   fun goBackToPreviousScreen()
   fun showSignatureRequiredError()
   fun openSharePrescriptionScreen(teleconsultRecordId: UUID, medicalInstructions: String)
+  fun showMedicinesRequiredError()
 }