+++ app/src/main/java/org/simple/clinic/teleconsultlog/prescription/medicines/TeleconsultMedicinesUi.kt
@@ -5,4 +5,7 @@ import org.simple.clinic.drugs.PrescribedDrug
 interface TeleconsultMedicinesUi {
   fun renderMedicines(medicines: List<PrescribedDrug>)
   fun showNoMedicines()
+  fun showAddButton()
+  fun showEditButton()
+  fun hideMedicinesRequiredError()
 }