+++ app/src/main/java/org/simple/clinic/teleconsultlog/prescription/medicines/TeleconsultMedicinesView.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import android.os.Parcelable
 import android.util.AttributeSet
 import android.view.LayoutInflater
+import android.view.View
 import androidx.appcompat.app.AppCompatActivity
 import androidx.constraintlayout.widget.ConstraintLayout
 import com.jakewharton.rxbinding3.view.clicks
@@ -131,6 +132,14 @@ class TeleconsultMedicinesView(
     medicinesRecyclerView.visibility = GONE
   }
 
+  override fun showAddButton() {
+    medicinesEditButton.text = context.getString(R.string.view_teleconsult_medicines_add)
+  }
+
+  override fun showEditButton() {
+    medicinesEditButton.text = context.getString(R.string.view_teleconsult_medicines_edit)
+  }
+
   override fun openEditMedicines(patientUuid: UUID) {
     screenRouter.push(PrescribedDrugsScreenKey(patientUuid))
   }
@@ -161,6 +170,14 @@ class TeleconsultMedicinesView(
         )
     )
     activity.startActivityForResult(intent, DRUG_FREQUENCY_SHEET)
+  }
+
+  fun showMedicinesRequiredError() {
+    medicinesRequiredErrorTextView.visibility = View.VISIBLE
+  }
+
+  override fun hideMedicinesRequiredError() {
+    medicinesRequiredErrorTextView.visibility = View.GONE
   }
 
   private fun editClicks(): Observable<UiEvent> {